+++ internal/config/flags.go
@@ -1,4 +1,4 @@
-package flags
+package config
 
 import (
 	"bytes"
@@ -34,14 +34,14 @@ type ParsedFlags struct {
 }
 
 // ParseFlags parses command-line arguments and returns the parsed flags.
-func ParseFlags(args []string, paramPrefix, version string, flagDocs map[string][]FlagDoc) (*ParsedFlags, error) {
+func ParseFlags(args []string, version string) (*ParsedFlags, error) {
 	var knownArgs []string
 	var dynamicArgs []string
 
 	// Separate known flags and dynamic target flags
 	for i := 0; i < len(args); i++ {
 		arg := args[i]
-		if !strings.HasPrefix(arg, fmt.Sprintf("--%s.", paramPrefix)) {
+		if !strings.HasPrefix(arg, fmt.Sprintf("--%s.", ParamPrefix)) {
 			knownArgs = append(knownArgs, arg)
 			continue
 		}
@@ -64,9 +64,9 @@ func ParseFlags(args []string, paramPrefix, version string, flagDocs map[string]
 
 	// Custom usage function
 	flagSet.Usage = func() {
-		fmt.Fprintf(&buf, "Usage: %s [OPTIONS] [--%s.<IDENTIFIER>.<PROPERTY> value]\n\nOptions:\n", flagSetName, paramPrefix)
+		fmt.Fprintf(&buf, "Usage: %s [OPTIONS] [--%s.<IDENTIFIER>.<PROPERTY> value]\n\nOptions:\n", flagSetName, ParamPrefix)
 		flagSet.PrintDefaults()
-		displayCheckerProperties(&buf, flagDocs)
+		displayCheckerProperties(&buf)
 	}
 
 	// Define known flags
