+++ internal/config/parser.go
@@ -1,4 +1,4 @@
-package parser
+package config
 
 import (
 	"fmt"
