+++ internal/config/utils.go
@@ -1,6 +1,4 @@
-package parser
-
-import "github.com/containeroo/portpatrol/internal/flags"
+package config
 
 // trackUnusedParams tracks parameters for validation.
 func trackUnusedParams(params map[string]string) map[string]struct{} {
@@ -18,12 +16,4 @@ func extractMapKeys(m map[string]struct{}) []string {
 		keys = append(keys, key)
 	}
 	return keys
-}
-
-func GenerateDocs() map[string][]flags.FlagDoc {
-	return map[string][]flags.FlagDoc{
-		"TCP Checker Properties":  tcpFlagDocs,
-		"HTTP Checker Properties": httpFlagDocs,
-		"ICMP Checker Properties": icmpFlagDocs,
-	}
 }
